push

+++ setup.py
@@ -11,7 +11,7 @@ from setup_util import clone_github_repo, cmd, install
 install("torch")
 import torch
 
-install("opencv-python")
+install("opencv-python", "cv2")
 for mod in 'fvcore iopath lpips datetime timm ftfy'.split():
     install(mod)
 install("pytorch-lightning")
@@ -38,8 +38,11 @@ def download_model(url, file_name = None):
     out_dir = DefaultPaths.model_path
     if file_name is None:
         file_name = url.split("/")[-1]
-    print(f"Downloading {url}")
-    cmd(f'curl -L --progress-bar -o {out_dir}/{file_name} {url}')
+    if path_exists(f"{out_dir}/{file_name}"):
+        print(f"{file_name} already exists")
+    else: 
+        print(f"Downloading {url}")
+        cmd(f'curl -L --progress-bar -o {out_dir}/{file_name} {url}')
 
 download_model("https://the-eye.eu/public/AI/models/512x512_diffusion_unconditional_ImageNet/512x512_diffusion_uncond_finetune_008100.pt")
 download_model("https://the-eye.eu/public/AI/models/v-diffusion/secondary_model_imagenet_2.pth")
